moved other main functionality into service
