+++ electron/preload/index.ts
@@ -9,7 +9,6 @@ import MarcherPage, {
     ModifiedMarcherPageArgs,
 } from "@/global/classes/MarcherPage";
 import { ModifiedPageArgs, NewPageArgs } from "@/global/classes/Page";
-import { TablesWithHistory } from "@/global/Constants";
 import { contextBridge, ipcRenderer, SaveDialogOptions } from "electron";
 import * as DbServices from "electron/database/database.services";
 import { DatabaseResponse } from "electron/database/DatabaseActions";
@@ -153,13 +152,9 @@ const APP_API = {
     databaseCreate: () => ipcRenderer.invoke("database:create"),
 
     // Triggers
-    onFetch: (callback: (type: (typeof TablesWithHistory)[number]) => void) =>
-        ipcRenderer.on("fetch:all", (event, type) => callback(type)),
-    removeFetchListener: () => ipcRenderer.removeAllListeners("fetch:all"),
-    sendSelectedPage: (selectedPageId: number) =>
-        ipcRenderer.send("send:selectedPage", selectedPageId),
-    sendSelectedMarchers: (selectedMarchersId: number[]) =>
-        ipcRenderer.send("send:selectedMarchers", selectedMarchersId),
+    // onFetch: (callback: (type: (typeof TablesWithHistory)[number]) => void) =>
+    //     ipcRenderer.on("fetch:all", (event, type) => callback(type)),
+    // removeFetchListener: () => ipcRenderer.removeAllListeners("fetch:all"),
     sendLockX: (lockX: boolean) =>
         ipcRenderer.invoke("settings:set", "lockMovementX", lockX),
     sendLockY: (lockY: boolean) =>
